pngs... so mangy pngs
made the icons, all of them. no exception. everything is made either by the pen tool on photoshop or the shapes.  photoshop source files are also added for further modification. I think it's safe to say after this edit we're done with the icons which were a big part of this project.